Add files via upload
Shifting the sonogram may shift the orientation of reconstruction artifact.  In this example, shifting the sonogram produces transaxial tomographic reconstructions with different orientation of band artifacts emanating from the high attenuation petrous pyramids of the skull.  By taking the median value of 3 images reconstructed from the same sonogram (but with shifts in the sonogram), the intrapetrous artifact is reduced.