Add files via upload
This python 3.6 program shows modification of the sinogram (obtained using the radon transform of a CT image available on the NIH NBIA website) to reduce reconstruction artifact from extremities included in a CT head image.  The improvement is small.  However, it does demonstrate potential improvements.  The sinogram in this demonstration was obtained using the radon transform of an already reconstructed image.  The original sinogram, obtained from the scan,  has the most information of the actual measurement.  The reconstruction of the axial image (or other image plane) processes the information so that we can visualize our concept of anatomy.  Artifacts arise during the measurement (motion, for example), and during reconstruction (incomplete models of beam hardening, filtering, etc.).  Processing the sinogram can reduce artifact.  Artifact reduction by processing the sinogram can potentially be better than processing the reconstructed image because of the greater information content of the sinogram.
I eventually want to try deep learning on the sinogram, not only for artifact reduction, but for identification of pathology (such as intracranial hemorrhage).